made newcollection component separate file, cleaned up comments and console logs in user collections, error message ternary in usercollections return
